pyproject.toml
[tool.poetry]
name = "slack-rag-bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
slack-bolt = "^1.20.1"
fastapi = "^0.115.0"
python-dotenv = "^1.0.1"
uvicorn = "^0.31.0"
aiohttp = "^3.10.6"
rq = "^1.16.2"
redis = "^5.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
